+++ libs/types/src/lib.rs
@@ -24,9 +24,6 @@ pub mod ids;
 pub mod investments;
 pub mod orders;
 pub mod permissions;
-pub mod pools;
-pub mod reserves;
 pub mod time;
 pub mod tokens;
-pub mod tranches;
 pub mod xcm;
