+++ pallets/loans/src/test_utils.rs
@@ -14,10 +14,7 @@
 //! Module provides testing utilities for benchmarking and tests.
 use cfg_primitives::CFG as CURRENCY;
 use cfg_traits::{Investment, PoolMutate, PoolNAV, TrancheCurrency as _};
-use cfg_types::{
-	tokens::{CurrencyId, TrancheCurrency},
-	tranches::{TrancheInput, TrancheLoc, TrancheMetadata, TrancheType},
-};
+use cfg_types::tokens::{CurrencyId, TrancheCurrency};
 use codec::Encode;
 use frame_support::{
 	assert_ok, parameter_types,
@@ -30,7 +27,10 @@ use frame_support::{
 };
 use frame_system::RawOrigin;
 #[cfg(feature = "runtime-benchmarks")]
-use pallet_pool_system::{Pallet as PoolPallet, Pool as PoolStorage};
+use pallet_pool_system::{
+	tranches::{TrancheInput, TrancheLoc, TrancheMetadata, TrancheType},
+	Pallet as PoolPallet, Pool as PoolStorage,
+};
 use sp_runtime::{
 	traits::{AccountIdConversion, Zero},
 	Perquintill,
