+++ pallets/loans/src/tests.rs
@@ -16,7 +16,6 @@ use cfg_primitives::{Balance, CollectionId, ItemId, PoolEpochId, PoolId, Tranche
 use cfg_traits::{InvestmentAccountant, InvestmentProperties};
 use cfg_types::{
 	fixed_point::Rate,
-	pools::PoolLocator,
 	tokens::{CurrencyId, TrancheCurrency},
 };
 use codec::MaxEncodedLen;
@@ -26,6 +25,7 @@ use frame_support::{
 };
 use loan_type::{BulletLoan, LoanType};
 use pallet_loans::Event as LoanEvent;
+use pallet_pool_system::pool_types::PoolLocator;
 use sp_arithmetic::{traits::checked_pow, FixedPointNumber};
 use sp_runtime::{
 	traits::{BadOrigin, MaybeSerializeDeserialize, StaticLookup},
