+++ pallets/pool-system/src/impls.rs
@@ -13,16 +13,14 @@
 use cfg_traits::{
 	CurrencyPair, InvestmentAccountant, PoolUpdateGuard, PriceValue, TrancheCurrency, UpdateState,
 };
-use cfg_types::{
-	epoch::{EpochState, ScheduledUpdateDetails},
-	investments::InvestmentInfo,
-	pools::{PoolDetails, PoolParameters, PoolStatus},
-	reserves::ReserveDetails,
-	tranches::{TrancheInput, TrancheLoc, TrancheUpdate, Tranches},
-};
+use cfg_types::{epoch::EpochState, investments::InvestmentInfo};
 use frame_support::traits::Contains;
 
 use super::*;
+use crate::{
+	pool_types::{PoolDetails, PoolParameters, PoolStatus, ReserveDetails, ScheduledUpdateDetails},
+	tranches::{TrancheInput, TrancheLoc, TrancheUpdate, Tranches},
+};
 
 impl<T: Config> PoolInspect<T::AccountId, T::CurrencyId> for Pallet<T> {
 	type Moment = Moment;
