+++ runtime/altair/src/lib.rs
@@ -25,14 +25,11 @@ use cfg_traits::{
 pub use cfg_types::tokens::CurrencyId;
 use cfg_types::{
 	consts::pools::*,
-	epoch::ScheduledUpdateDetails,
 	fee_keys::FeeKey,
 	fixed_point::Rate,
 	permissions::{PermissionRoles, PermissionScope, PermissionedCurrencyRole, PoolRole, Role},
-	pools::PoolDetails,
 	time::TimeProvider,
 	tokens::{CustomMetadata, TrancheCurrency},
-	tranches::{TrancheIndex, TrancheLoc, TrancheSolution},
 };
 use codec::{Decode, Encode, MaxEncodedLen};
 /// Constant values used within the runtime.
@@ -58,7 +55,11 @@ use pallet_anchors::AnchorData;
 pub use pallet_balances::Call as BalancesCall;
 use pallet_collective::{EnsureMember, EnsureProportionMoreThan};
 use pallet_investments::OrderType;
-use pallet_pool_system::EpochSolution;
+use pallet_pool_system::{
+	pool_types::{PoolDetails, ScheduledUpdateDetails},
+	tranches::{TrancheIndex, TrancheLoc, TrancheSolution},
+	EpochSolution,
+};
 use pallet_restricted_tokens::{FungibleInspectPassthrough, FungiblesInspectPassthrough};
 pub use pallet_timestamp::Call as TimestampCall;
 pub use pallet_transaction_payment::{CurrencyAdapter, Multiplier, TargetedFeeAdjustment};
