+++ runtime/integration-tests/src/xcm/development/tests/connectors.rs
@@ -32,7 +32,6 @@ use cfg_traits::PoolMutate;
 use cfg_types::{
 	fixed_point::Rate,
 	permissions::{PermissionScope, PoolRole, Role, UNION},
-	pools::PoolDetails,
 	tokens::{CurrencyId, CurrencyId::ForeignAsset, CustomMetadata, ForeignAssetId},
 	tranches::{TrancheInput, TrancheLoc, TrancheMetadata, TrancheType},
 	xcm::XcmMetadata,
@@ -49,6 +48,7 @@ use pallet_connectors::{
 	encoded_contract_call, Domain, DomainAddress, DomainLocator, Error::UnauthorizedTransfer,
 	Message, ParachainId, Router, XcmDomain,
 };
+use pallet_pool_system::pool_types::PoolDetails;
 use runtime_common::{xcm::general_key, xcm_fees::default_per_second};
 use sp_core::H160;
 use sp_runtime::{
