+++ src/rpc/pools.rs
@@ -1,9 +1,11 @@
 use std::{fmt::Debug, sync::Arc};
 
-use cfg_types::tranches::{TrancheIndex, TrancheLoc, TrancheSolution};
 use codec::Codec;
 use jsonrpsee::{core::RpcResult, proc_macros::rpc};
-use pallet_pool_system::EpochSolution;
+use pallet_pool_system::{
+	tranches::{TrancheIndex, TrancheLoc, TrancheSolution},
+	EpochSolution,
+};
 use runtime_common::apis::PoolsApi as PoolsRuntimeApi;
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
